Add files via upload
FLIGHT IMAGE CLASSIFICATION
“Predicting the future isn’t magic, it’s artificial
intelligence.”
~Dave Waters
ABSTRACT
This article's main goal is to emphasize the value of machine learning and artificial
intelligence in object detection. Sentiment analysis, which analyses and extracts
subjective information from the pertinent data, is nothing more than opinion mining
of text material. A computer vision technique called object detection is used to find
occurrences of objects in pictures or movies. Following the implementation of
several machine learning models, accuracy scores and performance metrics are
produced and compared. When classifying whether an object is a drone, a flight,
etc., the model that best fits the data or provides the highest accuracy score is taken
into account.
INTRODUCTION
To classify among the images of flights using a VGG-19 model trained on the A
CNN is a supervised learning method that requires the supply of input data as well
as goal output data. In order to give a trained model for next data analysis, these
are categorized using their labels.
Convolutional Layers, Pooling Layers, and Fully Connected Dense Networks are
the three basic components of a CNN in most cases. The input image is applied to
m number of n x n filters by the convolutional layer to produce a feature map. The
max pool layer, which is primarily utilized for dimensionality reduction and only
selects the best features from the feature map, is fed the feature map after that. The
fully connected dense neural network, which learns the weights through
backpropagation and outputs the classification, receives flattened versions of all
the features as input.
The motivation for the CNN is that it is based on how the visual cortex works,
where one thing in the view is in focus while the rest is blurred. In a similar
manner, the CNN classifies incoming images by taking a portion or window at a
time. Every time, the convolutional layer of the CNN will generate a feature map
for each region. It performs feature extraction in the Pooling layer by removing the
superfluous characteristics and retaining only the most crucial ones for that
section. Thus, by using CNNs, we may extract features without using a different
method.
PROBLEM STATEMENT
Flight Image Classification using Deep Learning Algorithms like CNN
Dataset Link:
https://drive.google.com/drive/folders/1vmetOF7T1ieJQEXwccrHmIiAXjP1dK mz

VGG16
A convolutional neural network is also known as a ConvNet, which is a kind of
artificial neural network. A convolutional neural network has an input layer, an
output layer, and various hidden layers. VGG16 is a type of CNN (Convolutional
Neural Network) that is considered to be one of the best computer vision models to
date. The creators of this model evaluated the networks and increased the depth
using an architecture with very small (3 × 3) convolution filters, which showed a
significant improvement on the prior-art configurations. They pushed the depth to
16–19 weight layers making it approximate 138 trainable parameters.
VGG16 is object detection and classification algorithm which is able to classify
1000 images of 1000 different categories with 96% accuracy. It is one of the
popular algorithms for image classification and is easy to use with transfer
learning.
VGG16 Architecture:
METHODOLOGY
We had a dataset of images having the categories like Helicopter, Missile,
Rocket, Drone, Fighter jet and Passenger plane. VGG-19 is one of the
pre-trained Convolutional Neural Networks (CNNs) and is 19 layers deep! Just
like other transfer learning models, it is trained on 1000 categories. Every time
we use it to classify a problem, we should customize the last layer of the model
according to the number of classes of our problem. I will elaborate practically on
this when we proceed to the section of building our model.
Step 1: Data cleaning
There were duplicate images and useless images in the dataset which are removed
by manually. There were some files whom extensions are not accepted by python
libraries. That files were removed manually.
Step-2: Pre-processing of the images
On our unique training dataset, we will train a VGG-19 model to distinguish
between the three categories of rock, paper, and scissors. An image in colour with
the dimensions 224224 of one of the three hand movements is fed into the
pre-trained CNN model. However, each image in the collection has a size of 300
by 200 pixels. As a result, they must all be adjusted to the necessary size.
Let's also generate x train, x test, x val, y train, y test, and y val while resizing the
photos. In a nutshell, these concepts are as follows:
x_train: Numpy arrays of the images of the training dataset
y_train: Labels of the training dataset
x_test: Numpy arrays of the images of the testing dataset
y_test: Labels of the testing dataset
x_val: Numpy arrays of the images of the validation dataset
y_val: Labels of the validation dataset
Step-3: Model training
Step-3.1: Model Building
As mentioned earlier, we will be using the VGG-19 pre-trained model to classify
rock, paper, and scissors. Thus, we are dealing with a multi-class classification
problem with 6 categories. We imported vgg-19 model. Now, to customize the
model, we changed its last layer alone according to the number of classes in our
problem.
Step-3.2: Compiling the model
We observed that the labels of none of the datasets are one-hot encoded. So, we
should use sparse categorical cross-entropy as our loss function. We will use the
best optimizer called Adam optimizer as it decides the best learning rate on its
own.
Step-3.3: Fitting the model
While training, we must watch out that our model does not get overly fit.
According to this phenomenon, our model will perform flawlessly on our train
dataset but will perform poorly on new data. In that situation, it is also referred to
as overfitting the model to the specific dataset.
Thus, we used early stopping to stop training the model any further if the validation
loss suddenly starts increasing.
Step-4: Model Evaluation
We evaluated our model by testing it on the test dataset. our deep learning model
good and efficient, it is not only enough to look at its accuracy but it is also
equally essential to observe its classification report and confusion matrix.
en.flow_from_directory(train_path,

OUTPUT
The first step is to build a Machine Learning Model to predict whether the given
input image is a flight, drone, missile, etc.
CONCEPT DEVELOPMENT
The first step is to build a Machine Learning Model to predict whether the given
input image is a flight, drone, missile, etc.
Now this model can be utilized effectively and can be used in:
● Computer vision
● Image recognition
● Natural language processing (NLP)
● Recommender systems
CONCLUSION
This project is aimed to predict whether the object image gets classified or
not. This CNN model will be the thing that clears up this stressful task. We
have built an efficient image classifier for our custom dataset. Thus, we have
trained a VGG-19 model to classify images. The accuracy which was
obtained was 90.8%. The model can further be trained for more epochs for
further better performance. We can extend the take-aways to build one such
image classifier for any other custom dataset as well.